# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Điện Biên Weather Alert API",
    "description": "\nAPI của **Hệ thống AI Cảnh báo & Dự báo Thời tiết Vi mô — Điện Biên**.\n\nCung cấp dữ liệu dự báo/cảnh báo thiên tai cấp huyện, bản tin đa ngôn ngữ (Việt · Thái · H'Mông),\nvà mô phỏng phân phối đa kênh cho giao diện người dân + dashboard chỉ huy.\n\n### Khái niệm chính\n- **Mức cảnh báo (4 cấp):** `Green` (Bình thường) · `Yellow` (Chú ý) · `Orange` (Nguy hiểm) · `Red` (Cực kỳ nguy hiểm).\n- **Khoá dữ liệu:** mỗi bản ghi định danh bằng `(location_id, date)` với `date` dạng `dd/mm/yyyy`.\n- **Audio:** các URL `/audio/...` trỏ tới file mp3 tĩnh (giọng đọc bản tin theo ngôn ngữ).\n\n### Ghi chú\n- API **chỉ đọc**; endpoint phân phối **chỉ mô phỏng**, không gửi SMS/Zalo/loa thật.\n- Tài liệu tương tác: **Swagger UI** tại `/docs`, **ReDoc** tại `/redoc`, đặc tả JSON tại `/openapi.json`.\n",
    "contact": {
      "name": "Đội dự thi — Vietnam AI Innovation Challenge"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/forecast/{location_id}": {
      "get": {
        "tags": [
          "Dự báo"
        ],
        "summary": "Dự báo 3–7 ngày của một huyện",
        "description": "Chuỗi dự báo theo ngày (đã sắp xếp tăng dần) cho một huyện, mỗi ngày là một bản ghi đã làm giàu.",
        "operationId": "get_forecast_api_forecast__location_id__get",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Mã huyện",
              "examples": [
                "muong_nhe"
              ],
              "title": "Location Id"
            },
            "description": "Mã huyện",
            "examples": {
              "muong_nhe": {
                "summary": "Huyện Mường Nhé",
                "value": "muong_nhe"
              },
              "muong_cha": {
                "summary": "Huyện Mường Chà",
                "value": "muong_cha"
              },
              "tuan_giao": {
                "summary": "Huyện Tuần Giáo",
                "value": "tuan_giao"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlertRecord"
                  },
                  "title": "Response Get Forecast Api Forecast  Location Id  Get"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy tài nguyên",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlertRecord": {
        "properties": {
          "location": {
            "type": "string",
            "title": "Location",
            "examples": [
              "Huyện Mường Nhé"
            ]
          },
          "location_id": {
            "type": "string",
            "title": "Location Id",
            "examples": [
              "muong_nhe"
            ]
          },
          "latitude": {
            "type": "number",
            "title": "Latitude",
            "examples": [
              22.1989
            ]
          },
          "longitude": {
            "type": "number",
            "title": "Longitude",
            "examples": [
              102.4481
            ]
          },
          "elevation": {
            "type": "integer",
            "title": "Elevation",
            "examples": [
              900
            ]
          },
          "date": {
            "type": "string",
            "title": "Date",
            "description": "Ngày dạng dd/mm/yyyy",
            "examples": [
              "20/07/2026"
            ]
          },
          "highest_alert_level": {
            "type": "string",
            "enum": [
              "Green",
              "Yellow",
              "Orange",
              "Red"
            ],
            "title": "Highest Alert Level",
            "description": "Mức cảnh báo cao nhất trong ngày"
          },
          "weather_summary": {
            "$ref": "#/components/schemas/WeatherSummary"
          },
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/HazardAlertOut"
            },
            "type": "array",
            "title": "Alerts",
            "description": "Danh sách hiểm họa được phát hiện"
          },
          "messages": {
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "enum": [
                "vi",
                "thai",
                "hmong"
              ]
            },
            "type": "object",
            "title": "Messages",
            "description": "Bản tin đa ngôn ngữ. Rỗng nếu ngày đó chưa có bản dịch (has_translation=false)."
          },
          "audio": {
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "enum": [
                "vi",
                "thai",
                "hmong"
              ]
            },
            "type": "object",
            "title": "Audio",
            "description": "URL file mp3 theo ngôn ngữ (chỉ gồm ngôn ngữ có audio thực tế). Phục vụ tại /audio/..."
          },
          "has_translation": {
            "type": "boolean",
            "title": "Has Translation",
            "description": "Ngày này đã có bản tin đa ngôn ngữ hay chưa"
          }
        },
        "type": "object",
        "required": [
          "location",
          "location_id",
          "latitude",
          "longitude",
          "elevation",
          "date",
          "highest_alert_level",
          "weather_summary",
          "alerts",
          "has_translation"
        ],
        "title": "AlertRecord",
        "example": {
          "alerts": [
            {
              "description": "Đất ngậm nước bão hòa kết hợp mưa lớn. Nguy cơ lũ quét và sạt lở núi cực cao!",
              "hazard": "Lũ quét & Sạt lở",
              "level": "Red"
            }
          ],
          "audio": {
            "vi": "/audio/Huy%E1%BB%87n_M%C6%B0%E1%BB%9Dng_Nh%C3%A9/20072026/vi.mp3"
          },
          "date": "20/07/2026",
          "elevation": 900,
          "has_translation": true,
          "highest_alert_level": "Red",
          "latitude": 22.1989,
          "location": "Huyện Mường Nhé",
          "location_id": "muong_nhe",
          "longitude": 102.4481,
          "messages": {
            "hmong": "...",
            "thai": "...",
            "vi": "Mường Nhé ngày 20/07/2026: Cảnh báo đỏ về lũ quét và sạt lở..."
          },
          "weather_summary": {
            "deep_soil_moisture": 0.42,
            "max_cape": 1620.0,
            "max_rain_1h": 12.0,
            "max_temp": 28.5,
            "max_wind_gust": 19.8,
            "min_temp": 21.0,
            "min_visibility": 800.0,
            "total_rain": 59.5
          }
        }
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Thông báo lỗi",
            "examples": [
              "Không tìm thấy địa điểm 'xyz'"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HazardAlertOut": {
        "properties": {
          "hazard": {
            "type": "string",
            "title": "Hazard",
            "description": "Tên nhóm hiểm họa",
            "examples": [
              "Lũ quét & Sạt lở"
            ]
          },
          "level": {
            "type": "string",
            "enum": [
              "Green",
              "Yellow",
              "Orange",
              "Red"
            ],
            "title": "Level",
            "description": "Mức cảnh báo của hiểm họa"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Mô tả chi tiết + khuyến cáo"
          }
        },
        "type": "object",
        "required": [
          "hazard",
          "level",
          "description"
        ],
        "title": "HazardAlertOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeatherSummary": {
        "properties": {
          "min_temp": {
            "type": "number",
            "title": "Min Temp",
            "description": "Nhiệt độ thấp nhất đã hiệu chỉnh (°C)",
            "examples": [
              21.0
            ]
          },
          "max_temp": {
            "type": "number",
            "title": "Max Temp",
            "description": "Nhiệt độ cao nhất đã hiệu chỉnh (°C)",
            "examples": [
              28.5
            ]
          },
          "total_rain": {
            "type": "number",
            "title": "Total Rain",
            "description": "Tổng lượng mưa 24h (mm)",
            "examples": [
              59.5
            ]
          },
          "max_rain_1h": {
            "type": "number",
            "title": "Max Rain 1H",
            "description": "Cường độ mưa lớn nhất trong 1h (mm)",
            "examples": [
              12.0
            ]
          },
          "max_wind_gust": {
            "type": "number",
            "title": "Max Wind Gust",
            "description": "Gió giật mạnh nhất (km/h)",
            "examples": [
              19.8
            ]
          },
          "max_cape": {
            "type": "number",
            "title": "Max Cape",
            "description": "CAPE lớn nhất (J/kg) — chỉ số đối lưu",
            "examples": [
              1620.0
            ]
          },
          "min_visibility": {
            "type": "number",
            "title": "Min Visibility",
            "description": "Tầm nhìn xa thấp nhất (m)",
            "examples": [
              800.0
            ]
          },
          "deep_soil_moisture": {
            "type": "number",
            "title": "Deep Soil Moisture",
            "description": "Độ ẩm đất tầng sâu 27–81cm (m³/m³)",
            "examples": [
              0.42
            ]
          }
        },
        "type": "object",
        "required": [
          "min_temp",
          "max_temp",
          "total_rain",
          "max_rain_1h",
          "max_wind_gust",
          "max_cape",
          "min_visibility",
          "deep_soil_moisture"
        ],
        "title": "WeatherSummary"
      },
...
    }
  },
...
}